refactor(review-pacing): extract the advanced-reveal chrome
scheduling-section.vue fused its field group with the reveal/scrim/badge
GSAP chrome. Split the reveal mechanics into a sibling advanced-reveal.vue
so scheduling-section is pure fields wrapped in the new component's slot.
